Score Value Display + Partial Credits
The score display for each selector button has been implemented. The displays update as the dice update, and once a selector is selected it will "lock" that selector's display for the rest of combat so it stops updating.

Additionally the credits menu is now partially implemented. The menu still needs to be graphically assembled. Additionally the buttons for going to and from the menu are bugged so they only sometimes work.