Credits + Menu Backdrop
Implemented the Credits option in the main menu. Works as specified in the "Credits" issue, with the added feature that clicking on the names in the menu/the GMTK Jam logo will open up relevant links.

Additionally added a rudimentary backdrop asset for the main menu so its not a completely blank void.